[change] フロントから来たdateをrailsで扱えるように変換する処理を追加

+++ app/controllers/api/v1/exam_controller.rb
@@ -92,24 +92,29 @@ class Api::V1::ExamController < ApplicationController
     end
 
     def update
-        ins = Exam.find_by(user_id:@user.id,id:params[:exam_id])
-        ins = ins.update(user_id:@user.id, title: params[:title], content: params[:content], examDate:params[:examdate], position:params[:position],complete:params[:complete])
-        if ins
-            render json: JSON.pretty_generate({
-                                              status:'SUCCESS',
-                                              api_version: 'v1',
-                                              mes:"試験を更新しました。"
-                                              })
-        else
-            render json: JSON.pretty_generate({
-                                                status:'Error',
-                                                api_version: 'v1',
-                                                mes: '試験の更新に失敗しました。'
-            
-                                                })
+        if @userSession
+            #sessionが有効だったらタスクを作る
+            ins = Exam.find_by(user_id:@user.id,id:params[:exam_id])
+            ins = ins.update(user_id:@user.id, title: params[:title], content: params[:content], examDate:params[:examdate], position:params[:position],complete:params[:complete])
+            if ins
+                render json: JSON.pretty_generate({
+                                                  status:'SUCCESS',
+                                                  api_version: 'v1',
+                                                  mes:"試験を更新しました。"
+                                                  })
+            else
+                render json: JSON.pretty_generate({
+                                                    status:'Error',
+                                                    api_version: 'v1',
+                                                    mes: '試験の更新に失敗しました。'
+                
+                                                    })
+            end
         end
     end
     def destroy
+        if @userSession
+        #sessionが有効だったらタスクを作る
           ins = Exam.find_by(user_id:@user.id,id:params[:exam_id])
           if ins.destroy
               render json: JSON.pretty_generate({
@@ -125,6 +130,6 @@ class Api::V1::ExamController < ApplicationController
               
                                                   })
           end
-        
+        end
     end
 end
