+++ app/controllers/api/v1/task_controller.rb
@@ -99,23 +99,28 @@ class Api::V1::TaskController < ApplicationController
         end
     end
     def update
-        ins = UserTask.find_by(user_id:@user.id,id:params[:task_id])
-        ins = ins.update(user_id:@user.id, title: params[:title], content: params[:content], taskDate:params[:taskdate], position:params[:position],complete:params[:complete])
-        if ins
-            render json: JSON.pretty_generate({
-                                              status:'SUCCESS',
-                                              api_version: 'v1',
-                                              mes:"タスクを更新しました。"
-                                              })
-        else
-            render json: JSON.pretty_generate({
-                                                status:'Error',
-                                                api_version: 'v1',
-                                                mes: 'タスクの更新に失敗しました。'
-                                                })
+        if @userSession
+            #sessionが有効だったらタスクを作る
+            ins = UserTask.find_by(user_id:@user.id,id:params[:task_id])
+            ins = ins.update(user_id:@user.id, title: params[:title], content: params[:content], taskDate:params[:taskdate], position:params[:position],complete:params[:complete])
+            if ins
+                render json: JSON.pretty_generate({
+                                                  status:'SUCCESS',
+                                                  api_version: 'v1',
+                                                  mes:"タスクを更新しました。"
+                                                  })
+            else
+                render json: JSON.pretty_generate({
+                                                    status:'Error',
+                                                    api_version: 'v1',
+                                                    mes: 'タスクの更新に失敗しました。'
+                                                    })
+            end
         end
     end
     def destroy
+        if @userSession
+        #sessionが有効だったらタスクを作る
           ins = UserTask.find_by(user_id:@user.id,id:params[:task_id])
           if ins.destroy
               render json: JSON.pretty_generate({
@@ -131,6 +136,6 @@ class Api::V1::TaskController < ApplicationController
               
                                                   })
           end
-        
+        end
     end
 end
